Edit index.html and add style.css.
Change the place of the equal button from ".controls"
to ".numbers" (index.html) .

Add link tag in index.html to link style.css file.

Apply css grid to the ".numbers" and all other
button to make it look like a calculator.

Adding margin to the body and padding to
".calculator_container" .